The most significant changes include the replacement of the namespace `HttpWebshopCookie.Utilities` with `HttpWebshopCookie.Config` in `ApplicationDbContext.cs`, the addition of a new configuration `SpecialOrderInstructionsConfiguration` to the `ApplicationDbContext` class, and the addition of a new static array `dateRange` to the `SeedAllData` class in `SeedAllData.cs`. The number of iterations in the for loops in the `SeedOrdersAsync` and `SeedBasketActivityAsync` methods of the `SeedAllData` class was increased from 10 to 20. The way order items are added to an order was changed from using `AddRange` to a foreach loop in the `SeedOrdersAsync` method of the `SeedAllData` class.
The `dateTime` variable and `orderDate` in the `SeedBasketActivityAsync` and `GenerateRandomDates` methods of the `SeedAllData` class were changed to use the `dateRange` array for generating random dates. The `Timestamp` in the `SeedBasketActivityAsync` method of the `SeedAllData` class was changed to add a random number of seconds between 10 and 30 instead of between 5 and 20.

The form layout and the layout of the cards in `CustomerInteraction.cshtml` and `ProductsOrders.cshtml` were updated to use Bootstrap's grid system and card classes. The `Back to Analytics` button in these files was changed to just `Back`. The chart options in these files were updated to include responsive settings and y-axis settings.

The `DateFrom` and `DateTo` properties in `CustomerInteraction.cshtml.cs` were updated to use the `BindProperty` attribute and the default values were changed from `DateTime.UtcNow` to `DateTime.Now`. The `OnGetAsync` method in `CustomerInteraction.cshtml.cs` was updated to call individual methods for fetching data instead of `FetchCustomerInteractionDataAsync`. The `FetchCustomerInteractionDataAsync` method was removed from `CustomerInteraction.cshtml.cs`.

The `FilterCustomersByDateRange` and `FilterBasketActivitiesByDateRange` methods in `CustomerInteraction.cshtml.cs` were updated to take date parameters instead of a query. The data fetching methods in `CustomerInteraction.cshtml.cs` were updated to take a query parameter instead of filtering the query within the method. The `AvgTimeSpentOnSite` calculation in `CustomerInteraction.cshtml.cs` was updated to return 0 if there are no session durations.

The `Period` and `NewCustomers` properties in the `CustomerGrowth` class, the `ActivityType` property in the `BasketActivitySummary` class, and the `ProductName` property in the `ProductActivity` class in `CustomerInteraction.cshtml.cs` were updated to be nullable.

The `ProductsOrdersModel` class in `ProductsOrders.cshtml.cs` file and the `SalesModel` class in `Sales.cshtml.cs` file were updated to include `BindProperty` attributes for `DateFrom` and `DateTo` properties. The `OnGetAsync` method in these files was updated to use these properties directly instead of taking them as parameters.

The `OrderListings.cshtml` file was updated to change the `asp-page` attribute of the `a` element in the last `td` element to use a relative path. The `_Layout.cshtml` file was updated to include more versions of Chart.js and to include the site's own JavaScript files. The `Program.cs` file was updated to remove the global using directive for `NuGet.Packaging`.