`Remove product table from _BasketPartial.cshtml`
Removed the entire code block in the `_BasketPartial.cshtml` file. This code block was responsible for creating a table that displayed a list of products with their name, price, description, and quantity. It also provided links to edit, view details, or delete each product, as well as options to add or remove the product from the basket.