Update button attributes and icon in UI
Updated the `asp-page` attribute in the button element to navigate to the Edit page instead of the Details page. The `title` attribute of the button has been changed to "Edit Product" to reflect the new functionality. The button's icon has also been updated from a minus circle to a magnifying glass to visually indicate the new function.