Added HttpWebshopCookie.Tests project and refactored BasketService for testability
A new test project `HttpWebshopCookie.Tests` has been added to the solution, with several test classes for the service classes in the `HttpWebshopCookie` project. The `BasketService` class has been refactored to make it more testable, with overloaded methods to allow mocking of dependencies. New package references for integration testing and in-memory database testing have been added. The `launchSettings.json` file has been updated to include a profile for running the new test project. A `TestStartup` class has been added to set up the application for testing with an in-memory database.