Uncommented unit test in BasketServiceTests.cs
Uncommented the `GetOrCreateBasket_ShouldCreateNewBasketIfNotExists()` method in the `BasketServiceTests.cs` file. This unit test checks if a new basket is created when it does not exist, following the Arrange-Act-Assert pattern. The Arrange step creates a `BasketService` instance and appends a cookie named "BasketId" to the HTTP response. The Act step calls the `GetOrCreateBasket` method of the `BasketService` instance. The Assert step verifies the behavior of the `GetOrCreateBasket` method, checking if the basket is not null, if the count of baskets in the context is 1, if the "BasketId" cookie does not exist in the HTTP request, and if the cookie value is null or empty.