`Added <PublishTrimmed> and <StartupObject> to HttpWebshopCookie.csproj`
` `

`Added the <PublishTrimmed> property to the <PropertyGroup> section of the HttpWebshopCookie.csproj file and set it to False. This means the published output will not be trimmed to remove unused assemblies, potentially increasing the size of the published output. Also added an empty <StartupObject> property to the <PropertyGroup> section. This property is typically used to specify the entry point of the application when it's a console application. In this case, it doesn't change the behavior of the project as it's left empty.`