`Update JSON file configuration in Program.cs`
` `

`Updated the optional status of `appsettings.{EnvironmentName}.json` and `secrets.json` files from `false` to `true` in the `Program.cs` file. This change allows the application to run even if these files are missing. The `reloadOnChange` parameter remains `true` for both files, enabling the application to reload the configuration if these files are changed during runtime.