`Update SmtpClient connection method in EmailService`
Updated the connection method of the `SmtpClient` in the `EmailService` class. The previous `MailKit.Security.SecureSocketOptions.Auto` option has been replaced with the `MailKit.Security.SecureSocketOptions.StartTlsWhenAvailable` option. This change allows the connection to continue even if the server does not support STARTTLS. The modification was made in two places within the `EmailService` class.